Make CUDA an extension
Apply formatter
