+++ Project.toml
@@ -7,7 +7,6 @@ version = "0.13.4"
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 BandedMatrices = "aae01518-5342-5314-be14-df237901396f"
 BlockArrays = "8e7c35d0-a365-5155-bbbb-fb81a777f24e"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 ClimaComms = "3a4d1b5c-c61d-41fd-a00a-5873ba7a1b0d"
 CubedSphere = "7445602f-e544-4518-8976-18f8e8ae6cdb"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
@@ -32,9 +31,11 @@ Unrolled = "9602ed7d-8fef-5bc8-8597-8f21381861e8"
 
 [weakdeps]
 Krylov = "ba0b0d4f-ebba-5204-a429-3ac8c609bfb7"
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 
 [extensions]
 KrylovExt = "Krylov"
+ClimaCoreCUDAExt = "CUDA"
 
 [compat]
 Adapt = "3, 4"
