+++ ext/cuda/operators_integral.jl
@@ -1,0 +1,87 @@
+import ClimaCore: Fields, Spaces, Topologies
+import ClimaComms
+using CUDA: @cuda
+
+function column_integral_definite!(
+    ::ClimaComms.CUDADevice,
+    ∫field::Fields.Field,
+    ᶜfield::Fields.Field,
+)
+    space = axes(∫field)
+    Ni, Nj, _, _, Nh = size(Fields.field_values(∫field))
+    nthreads, nblocks = Topologies._configure_threadblock(Ni * Nj * Nh)
+    @cuda always_inline = true threads = nthreads blocks = nblocks column_integral_definite_kernel!(
+        strip_space(∫field, space),
+        strip_space(ᶜfield, space),
+    )
+end
+
+function column_integral_definite_kernel!(
+    ∫field,
+    ᶜfield::Fields.CenterExtrudedFiniteDifferenceField,
+)
+    idx = threadIdx().x + (blockIdx().x - 1) * blockDim().x
+    Ni, Nj, _, _, Nh = size(Fields.field_values(ᶜfield))
+    if idx <= Ni * Nj * Nh
+        i, j, h = Topologies._get_idx((Ni, Nj, Nh), idx)
+        ∫field_column = Spaces.column(∫field, i, j, h)
+        ᶜfield_column = Spaces.column(ᶜfield, i, j, h)
+        _column_integral_definite!(∫field_column, ᶜfield_column)
+    end
+    return nothing
+end
+
+
+function column_integral_indefinite!(
+    ::ClimaComms.CUDADevice,
+    ᶠ∫field::Fields.Field,
+    ᶜfield::Fields.Field,
+)
+    Ni, Nj, _, _, Nh = size(Fields.field_values(ᶠ∫field))
+    nthreads, nblocks = Topologies._configure_threadblock(Ni * Nj * Nh)
+    @cuda always_inline = true threads = nthreads blocks = nblocks column_integral_indefinite_kernel!(
+        ᶠ∫field,
+        ᶜfield,
+    )
+end
+
+function column_mapreduce_device!(
+    ::ClimaComms.CUDADevice,
+    fn::F,
+    op::O,
+    reduced_field::Fields.Field,
+    fields::Fields.Field...,
+) where {F, O}
+    Ni, Nj, _, _, Nh = size(Fields.field_values(reduced_field))
+    nthreads, nblocks = Topologies._configure_threadblock(Ni * Nj * Nh)
+    kernel! = if first(fields) isa Fields.ExtrudedFiniteDifferenceField
+        column_mapreduce_kernel_extruded!
+    else
+        column_mapreduce_kernel!
+    end
+    @cuda always_inline = true threads = nthreads blocks = nblocks kernel!(
+        fn,
+        op,
+        # reduced_field,
+        strip_space(reduced_field, axes(reduced_field)),
+        # fields...,
+        map(field -> strip_space(field, axes(field)), fields)...,
+    )
+end
+
+function column_mapreduce_kernel_extruded!(
+    fn::F,
+    op::O,
+    reduced_field,
+    fields...,
+) where {F, O}
+    idx = threadIdx().x + (blockIdx().x - 1) * blockDim().x
+    Ni, Nj, _, _, Nh = size(Fields.field_values(reduced_field))
+    if idx <= Ni * Nj * Nh
+        i, j, h = Topologies._get_idx((Ni, Nj, Nh), idx)
+        reduced_field_column = Spaces.column(reduced_field, i, j, h)
+        field_columns = map(field -> Spaces.column(field, i, j, h), fields)
+        _column_mapreduce!(fn, op, reduced_field_column, field_columns...)
+    end
+    return nothing
+end
