+++ ext/cuda/spaces.jl
@@ -1,0 +1,6 @@
+import ClimaCore: Spaces, AbstractSpace, bycolumn
+import ClimaComms
+function bycolumn(fn, space::AbstractSpace, ::ClimaComms.CUDADevice)
+    fn(:)
+    return nothing
+end
