+++ ext/cuda/topologies_dss.jl
@@ -1,4 +1,5 @@
-
+import ClimaCore: DataLayouts, Topologies, Spaces, Fields
+import ClimaCore.Topologies: DSSBuffer, Perimeter2D
 _max_threads_cuda() = 256
 
 function _configure_threadblock(max_threads, nitems)
