+++ src/DataLayouts/DataLayouts.jl
@@ -17,6 +17,7 @@ module DataLayouts
 import Base: Base, @propagate_inbounds
 import StaticArrays: SOneTo, MArray, SArray
 import ClimaComms
+import Adapt
 
 import ..slab, ..slab_args, ..column, ..column_args, ..level
 export slab, column, level, IJFH, IJF, IFH, IF, VF, VIJFH, VIFH, DataF
@@ -1422,7 +1423,32 @@ rebuild(data::AbstractData, ::Type{DA}) where {DA} =
 # broadcast machinery
 include("broadcast.jl")
 
-# GPU method specializations
-include("cuda.jl")
+
+Adapt.adapt_structure(to, data::IJKFVH{S, Nij, Nk}) where {S, Nij, Nk} =
+    IJKFVH{S, Nij, Nk}(Adapt.adapt(to, parent(data)))
+
+Adapt.adapt_structure(to, data::IJFH{S, Nij}) where {S, Nij} =
+    IJFH{S, Nij}(Adapt.adapt(to, parent(data)))
+
+Adapt.adapt_structure(to, data::VIJFH{S, Nij}) where {S, Nij} =
+    VIJFH{S, Nij}(Adapt.adapt(to, parent(data)))
+
+Adapt.adapt_structure(to, data::VIFH{S, Ni, A}) where {S, Ni, A} =
+    VIFH{S, Ni}(Adapt.adapt(to, parent(data)))
+
+Adapt.adapt_structure(to, data::IFH{S, Ni}) where {S, Ni} =
+    IFH{S, Ni}(Adapt.adapt(to, parent(data)))
+
+Adapt.adapt_structure(to, data::IJF{S, Nij}) where {S, Nij} =
+    IJF{S, Nij}(Adapt.adapt(to, parent(data)))
+
+Adapt.adapt_structure(to, data::IF{S, Ni}) where {S, Ni} =
+    IF{S, Ni}(Adapt.adapt(to, parent(data)))
+
+Adapt.adapt_structure(to, data::VF{S}) where {S} =
+    VF{S}(Adapt.adapt(to, parent(data)))
+
+Adapt.adapt_structure(to, data::DataF{S}) where {S} =
+    DataF{S}(Adapt.adapt(to, parent(data)))
 
 end # module
