+++ src/Fields/broadcast.jl
@@ -35,28 +35,6 @@ Base.Broadcast.BroadcastStyle(
 
 Base.Broadcast.broadcastable(field::Field) = field
 
-function Adapt.adapt_structure(
-    to::CUDA.KernelAdaptor,
-    bc::Base.Broadcast.Broadcasted{Style},
-) where {Style <: AbstractFieldStyle}
-    Base.Broadcast.Broadcasted{Style}(
-        Adapt.adapt(to, bc.f),
-        Adapt.adapt(to, bc.args),
-        Adapt.adapt(to, bc.axes),
-    )
-end
-
-function Adapt.adapt_structure(
-    to::CUDA.KernelAdaptor,
-    bc::Base.Broadcast.Broadcasted{Style, <:Any, Type{T}},
-) where {Style <: AbstractFieldStyle, T}
-    Base.Broadcast.Broadcasted{Style}(
-        (x...) -> T(x...),
-        Adapt.adapt(to, bc.args),
-        bc.axes,
-    )
-end
-
 Base.eltype(bc::Base.Broadcast.Broadcasted{<:AbstractFieldStyle}) =
     Base.Broadcast.combine_eltypes(bc.f, bc.args)
