+++ src/MatrixFields/field2arrays.jl
@@ -53,13 +53,15 @@ function column_field2array(field::Fields.FiniteDifferenceField)
             last_row = matrix_d < n_cols - n_rows ? n_rows : n_cols - matrix_d
 
             diagonal_data_view = view(diagonal_data, first_row:last_row)
-            CUDA.@allowscalar copyto!(matrix_diagonal, diagonal_data_view)
+            allow_scalar_func(copyto!, (matrix_diagonal, diagonal_data_view))
         end
         return matrix
     else # field represents a vector
-        return CUDA.@allowscalar Array(column_field2array_view(field))
+        return allow_scalar_func(Array, (column_field2array_view(field),))
     end
 end
+
+allow_scalar_func(::ClimaComms.AbstractDevice, f, args) = f(args...)
 
 """
     column_field2array_view(field)
