+++ src/MatrixFields/field_matrix_iterative_solver.jl
@@ -427,9 +427,11 @@ end
 check_field_matrix_solver(alg::StationaryIterativeSolve, cache, A, b) =
     check_preconditioner(alg.P_alg, cache.P_cache, A, b)
 
+is_CuArray_type(::Type{T}) where {T} = false
 function run_field_matrix_solver!(alg::StationaryIterativeSolve, cache, x, A, b)
     P = lazy_or_concrete_preconditioner(alg.P_alg, cache.P_cache, A)
-    using_cuda = ClimaComms.array_type(concrete_field_vector(b)) <: CUDA.CuArray
+    using_cuda =
+        is_CuArray_type(ClimaComms.array_type(concrete_field_vector(b)))
     if get_debug(alg) && !using_cuda
         @debug begin
             e₀ = concrete_field_vector(b) # Initialize e to any nonzero vector.
