+++ src/Operators/integrals.jl
@@ -15,35 +15,6 @@ horizontal space.
 column_integral_definite!(∫field::Fields.Field, ᶜfield::Fields.Field) =
     column_integral_definite!(ClimaComms.device(axes(ᶜfield)), ∫field, ᶜfield)
 
-function column_integral_definite!(
-    ::ClimaComms.CUDADevice,
-    ∫field::Fields.Field,
-    ᶜfield::Fields.Field,
-)
-    space = axes(∫field)
-    Ni, Nj, _, _, Nh = size(Fields.field_values(∫field))
-    nthreads, nblocks = Topologies._configure_threadblock(Ni * Nj * Nh)
-    @cuda always_inline = true threads = nthreads blocks = nblocks column_integral_definite_kernel!(
-        strip_space(∫field, space),
-        strip_space(ᶜfield, space),
-    )
-end
-
-function column_integral_definite_kernel!(
-    ∫field,
-    ᶜfield::Fields.CenterExtrudedFiniteDifferenceField,
-)
-    idx = threadIdx().x + (blockIdx().x - 1) * blockDim().x
-    Ni, Nj, _, _, Nh = size(Fields.field_values(ᶜfield))
-    if idx <= Ni * Nj * Nh
-        i, j, h = Topologies._get_idx((Ni, Nj, Nh), idx)
-        ∫field_column = Spaces.column(∫field, i, j, h)
-        ᶜfield_column = Spaces.column(ᶜfield, i, j, h)
-        _column_integral_definite!(∫field_column, ᶜfield_column)
-    end
-    return nothing
-end
-
 function column_integral_definite_kernel!(
     ∫field,
     ᶜfield::Fields.CenterFiniteDifferenceField,
@@ -106,19 +77,6 @@ The indefinite integral `ᶠ∫field = ϕ(z) = ∫ f(ϕ,z) dz` given:
 """
 column_integral_indefinite!(ᶠ∫field::Fields.Field, ᶜfield::Fields.Field) =
     column_integral_indefinite!(ClimaComms.device(ᶠ∫field), ᶠ∫field, ᶜfield)
-
-function column_integral_indefinite!(
-    ::ClimaComms.CUDADevice,
-    ᶠ∫field::Fields.Field,
-    ᶜfield::Fields.Field,
-)
-    Ni, Nj, _, _, Nh = size(Fields.field_values(ᶠ∫field))
-    nthreads, nblocks = Topologies._configure_threadblock(Ni * Nj * Nh)
-    @cuda always_inline = true threads = nthreads blocks = nblocks column_integral_indefinite_kernel!(
-        ᶠ∫field,
-        ᶜfield,
-    )
-end
 
 function column_integral_indefinite_kernel!(
     ᶠ∫field::Fields.FaceExtrudedFiniteDifferenceField,
@@ -280,47 +238,6 @@ column_mapreduce!(
     reduced_field,
     fields...,
 )
-
-function column_mapreduce_device!(
-    ::ClimaComms.CUDADevice,
-    fn::F,
-    op::O,
-    reduced_field::Fields.Field,
-    fields::Fields.Field...,
-) where {F, O}
-    Ni, Nj, _, _, Nh = size(Fields.field_values(reduced_field))
-    nthreads, nblocks = Topologies._configure_threadblock(Ni * Nj * Nh)
-    kernel! = if first(fields) isa Fields.ExtrudedFiniteDifferenceField
-        column_mapreduce_kernel_extruded!
-    else
-        column_mapreduce_kernel!
-    end
-    @cuda always_inline = true threads = nthreads blocks = nblocks kernel!(
-        fn,
-        op,
-        # reduced_field,
-        strip_space(reduced_field, axes(reduced_field)),
-        # fields...,
-        map(field -> strip_space(field, axes(field)), fields)...,
-    )
-end
-
-function column_mapreduce_kernel_extruded!(
-    fn::F,
-    op::O,
-    reduced_field,
-    fields...,
-) where {F, O}
-    idx = threadIdx().x + (blockIdx().x - 1) * blockDim().x
-    Ni, Nj, _, _, Nh = size(Fields.field_values(reduced_field))
-    if idx <= Ni * Nj * Nh
-        i, j, h = Topologies._get_idx((Ni, Nj, Nh), idx)
-        reduced_field_column = Spaces.column(reduced_field, i, j, h)
-        field_columns = map(field -> Spaces.column(field, i, j, h), fields)
-        _column_mapreduce!(fn, op, reduced_field_column, field_columns...)
-    end
-    return nothing
-end
 
 column_mapreduce_kernel!(fn::F, op::O, reduced_field, fields...) where {F, O} =
     _column_mapreduce!(fn, op, reduced_field, fields...)
